refactor(push): add a message policy class
